Created the Scraper
I had an idea for creating an automatic schedule creator way back when I first enrolled in college, but I didn't have the technical know-how to get it done. I still don't, but I tried this anyways. Scraping the website was the easy part; creating the algorithm to create automatic schedules is the difficult part. At this point, the scraper is entirely done, but the logic behind the schedule creator needs to be tweaked.