Updated Function
The app works now! Or, at least, it now works with a new purpose; previously, the app only searched for courses online. Now, the app can build a schedule using those classes scraped from the website. It was extremely difficult. I still need to add a function to add a range for hours between classes, and a function to make schedules with online classes in them. Currently, the schedules that are made have ONLY in-person or hybrid courses.