Fixed Counter
I was finally able to find out what was going wrong with the counter. It turns out, a method that was supposed to check each course timeslot for compatibility with each other never ran, and the "isWithinConstraints" method ran instead, incrementing the total possible courses int an unwanted amount of times. Now, the counter is incremented properly. Also, I reduced the amount of loops required by the for loop by adding a check for if the list has an online class within it every recurse.