Finally Finished the EXE
I had to use IntelliJ, build the artifact through that, then use Launch4J. The problem was that the installer didn't install properly, and the use of Launch4J completely mitigates that. I still don't know why the installer I created with jpackage wouldn't work, but that's a problem for another day. The app works as an exe now, and that's all that matters.