Changed Radio Buttons
The old radio buttons for the semester would return values that were finicky and didn't reflect well the semesters that they were actually representing. Also, the year spinner was useless. Instead, I added the website's text onto the radio buttons to better reflect their semester values. This way, it'll always be up-to-date.